+++ src/app/page.tsx
@@ -1,59 +1,33 @@
 "use client"
 
+import { Authenticated, Unauthenticated, AuthLoading } from "convex/react";
 import { NewToDoForm } from "./_components/new-todo-form";
-import { useQuery, useMutation } from "convex/react";
-import { api } from "../../convex/_generated/api";
-import { Id } from "../../convex/_generated/dataModel";
+import { ToDoList } from "./_components/to-do-list";
+import { SignInButton, UserButton } from "@clerk/nextjs";
 
 export default function Home() {
-  const todos = useQuery(api.functions.listTodos);
   return (
     <div className="max-w-screen-md mx-auto p-4 space-y-4">
-      <h1 className="text-xl font-bold">
-        Todo List
-      </h1>
-      <ul className="space-y-2">
-      {todos?.map(({_id, title, description, completed}, index) => (
-        <ToDoItem 
-          id={_id}
-          key={index}
-          title ={title} 
-          description ={description} 
-          completed ={completed}
-        />
-      ))}
-      </ul>
-      <NewToDoForm/>
+      <Authenticated>
+        <div className="div flex items-center justify-between">
+          <h1 className="text-xl font-bold">Welcome to Todo List Application</h1>
+          <UserButton/>
+        </div>
+        <ToDoList/>
+        <NewToDoForm/>
+      </Authenticated>
+
+      <Unauthenticated>
+        <p className="text-gray-700">Please Sign In to Continue</p>
+        <SignInButton>
+          <button className="p-1 bg-blue-600 text-white rounded">Sign In</button>
+        </SignInButton>
+      </Unauthenticated>
+      
+      <AuthLoading>
+        <p>Loading... Wait One Second!</p>
+      </AuthLoading>
      
     </div>
   );
-}
-
-function ToDoItem({id, title, description, completed}: {
-  id: Id<"todos">;
-  title: string; 
-  description: string; 
-  completed:boolean; 
-}) {
-
-  // methods from the functions.ts file
-  const updateTodo = useMutation(api.functions.updateTodo);
-  const deleteTodo = useMutation(api.functions.deleteTodo);
-
-  return (
-    <li className="w-full flex items-center gap-2 border rounded p-2">
-          <input 
-          type="checkbox" 
-          checked={completed} 
-          onChange={e => updateTodo({id, completed: e.target.checked})}
-        />
-        <div>
-          <p className="font-semibold">{title}</p>
-          <p className="text-sm text-gray-600">{description}</p>
-        </div>
-        <div className="ml-auto">
-          <button type="button" className="text-red-500" onClick={() => deleteTodo({id})}>Remove</button>
-        </div>
-        </li>
-  )
 }
